// lib/regex.d.ts
/**
 * @description These rules translate from the singular form of a noun to its plural form.
 * @private
 */
export declare const regex: {
    plural: {
        men: RegExp;
        people: RegExp;
        children: RegExp;
        tia: RegExp;
        analyses: RegExp;
        hives: RegExp;
        curves: RegExp;
        lrves: RegExp;
        aves: RegExp;
        foves: RegExp;
        movies: RegExp;
        aeiouyies: RegExp;
        series: RegExp;
        xes: RegExp;
        mice: RegExp;
        buses: RegExp;
        oes: RegExp;
        shoes: RegExp;
        crises: RegExp;
        octopi: RegExp;
        aliases: RegExp;
        summonses: RegExp;
        oxen: RegExp;
        matrices: RegExp;
        vertices: RegExp;
        feet: RegExp;
        teeth: RegExp;
        geese: RegExp;
        quizzes: RegExp;
        whereases: RegExp;
        criteria: RegExp;
        genera: RegExp;
        ss: RegExp;
        s: RegExp;
    };
    singular: {
        man: RegExp;
        person: RegExp;
        child: RegExp;
        ox: RegExp;
        axis: RegExp;
        octopus: RegExp;
        alias: RegExp;
        summons: RegExp;
        bus: RegExp;
        buffalo: RegExp;
        tium: RegExp;
        sis: RegExp;
        ffe: RegExp;
        hive: RegExp;
        aeiouyy: RegExp;
        x: RegExp;
        matrix: RegExp;
        vertex: RegExp;
        mouse: RegExp;
        foot: RegExp;
        tooth: RegExp;
        goose: RegExp;
        quiz: RegExp;
        whereas: RegExp;
        criterion: RegExp;
        genus: RegExp;
        s: RegExp;
        common: RegExp;
    };
};
export default regex;